pyproject.toml
[project]
name = "kagent"
version = "0.0.2"
description = "A wrapper around LangGraph, LangChain's callback handler, and WebSockets."
requires-python = ">=3.11,<4.0.0"
dependencies = [
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.0",
    "langchain-postgres>=0.0.12",
    "langchain>=0.3.14",
    "langgraph-checkpoint-postgres>=2.0.13",
    "langgraph>=0.2.62",
    "langsmith>=0.2.10",
    "python-socketio>=5.12.1",
    "psycopg-binary>=3.2.4",
    "psycopg-pool>=3.2.4",
    "psycopg>=3.2.4"
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["kagent", "kagent.nodes"]
package-dir = { "kagent" = "src" }

[tool.setuptools.package-data]
kagent = ["*"]
